act_2_decouple/requirements.txt
mysql-connector-python==8.0.31
protobuf==3.20.1
python-decouple==3.6
python-dotenv==0.21.0
